+++ Projects/4.Native-DeliveriesApp/DeliveriesApp/DeliveriesApp.Droid/DeliveriesApp.Droid.csproj
@@ -64,7 +64,6 @@
     <Compile Include="RegisterActivity.cs" />
     <Compile Include="Resources\Resource.designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Users.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Resources\AboutResources.txt" />
