+++ Projects/4.Native-DeliveriesApp/DeliveriesApp/DeliveriesApp.Droid/MainActivity.cs
@@ -6,6 +6,7 @@ using Android.OS;
 using Android.Support.V7.App;
 using Android.Runtime;
 using Android.Widget;
+using DeliveriesApp.Model;
 using Microsoft.WindowsAzure.MobileServices;
 
 namespace DeliveriesApp.Droid
@@ -13,8 +14,6 @@ namespace DeliveriesApp.Droid
     [Activity(Label = "DeliveriesApp.Droid", Theme = "@style/AppTheme", MainLauncher = true)]
     public class MainActivity : AppCompatActivity
     {
-        public static MobileServiceClient MobileServiceClient = new MobileServiceClient("https://xamarindeliveriesappglareone.azurewebsites.net", new HttpClientHandler());
-
         private EditText _emailEditText;
         private EditText _passwordEditText;
 
@@ -66,10 +65,9 @@ namespace DeliveriesApp.Droid
                 return;
             }
 
-            var user = (await MobileServiceClient.GetTable<Users>().Where(u => u.Email == email && u.Password == password)
-                .ToListAsync()).FirstOrDefault();
+            var loginResult = await Users.Login(email, password);
 
-            if (user == null)
+            if (!loginResult)
             {
                 Toast.MakeText(this, "Email and password are incorrect", ToastLength.Long).Show();
                 return;
