--- Projects/4.Native-DeliveriesApp/DeliveriesApp/DeliveriesApp.Droid/Users.cs
@@ -1,9 +1,0 @@
-﻿namespace DeliveriesApp.Droid
-{
-    public class Users
-    {
-        public string Id { get; set; }
-        public string Email { get; set; }
-        public string Password { get; set; }
-    }
-}
