+++ Projects/4.Native-DeliveriesApp/DeliveriesApp/DeliveriesApp.IOS/DeliveriesApp.IOS.csproj
@@ -145,7 +145,6 @@
     <Compile Include="RegisterViewController.designer.cs">
       <DependentUpon>RegisterViewController.cs</DependentUpon>
     </Compile>
-    <Compile Include="User.cs" />
     <Compile Include="ViewController.cs" />
     <Compile Include="ViewController.designer.cs">
       <DependentUpon>ViewController.cs</DependentUpon>
