--- Projects/4.Native-DeliveriesApp/DeliveriesApp/DeliveriesApp.IOS/User.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-using Foundation;
-using UIKit;
-
-namespace DeliveriesApp.iOS
-{
-    public class User
-    {
-        public string Id { get; set; }
-        public string Email { get; set; }
-        public string Password { get; set; }
-    }
-}
