--- Projects/4.Native-DeliveriesApp/DeliveriesApp/DeliveriesApp/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace DeliveriesApp
-{
-    public class Class1
-    {
-    }
-}
